Consolidate all server chat hooks in serverchathooks.cpp
