+++ NorthstarDedicatedTest/serverchathooks.cpp
@@ -3,6 +3,9 @@
 #include <rapidjson/document.h>
 #include <rapidjson/stringbuffer.h>
 #include <rapidjson/writer.h>
+#include "serverauthentication.h"
+#include "squirrel.h"
+#include "miscserverscript.h"
 
 class CServerGameDLL;
 class CBasePlayer;
@@ -50,17 +53,87 @@ MessageWriteStringType MessageWriteString;
 typedef void(__fastcall* MessageWriteBoolType)(bool bValue);
 MessageWriteBoolType MessageWriteBool;
 
-bool gSkipSayTextHook = false;
+static bool g_SkipSayTextHook = false;
+
+static std::string currentMessage;
+static int currentPlayerId;
+static int currentChannelId;
+static bool shouldBlock;
+static bool isProcessed = true;
+
+static SQRESULT setMessage(void* sqvm)
+{
+	currentMessage = ServerSq_getstring(sqvm, 1);
+	currentPlayerId = ServerSq_getinteger(sqvm, 2);
+	currentChannelId = ServerSq_getinteger(sqvm, 3);
+	shouldBlock = ServerSq_getbool(sqvm, 4);
+	return SQRESULT_NOTNULL;
+}
+static SQRESULT getMessageServer(void* sqvm)
+{
+	ServerSq_pushstring(sqvm, currentMessage.c_str(), -1);
+	return SQRESULT_NOTNULL;
+}
+static SQRESULT getPlayerServer(void* sqvm)
+{
+	ServerSq_pushinteger(sqvm, currentPlayerId);
+	return SQRESULT_NOTNULL;
+}
+static SQRESULT getChannelServer(void* sqvm)
+{
+	ServerSq_pushinteger(sqvm, currentChannelId);
+	return SQRESULT_NOTNULL;
+}
+static SQRESULT getShouldProcessMessage(void* sqvm)
+{
+	ServerSq_pushbool(sqvm, !isProcessed);
+	return SQRESULT_NOTNULL;
+}
+static SQRESULT pushMessage(void* sqvm)
+{
+	currentMessage = ServerSq_getstring(sqvm, 1);
+	currentPlayerId = ServerSq_getinteger(sqvm, 2);
+	currentChannelId = ServerSq_getinteger(sqvm, 3);
+	shouldBlock = ServerSq_getbool(sqvm, 4);
+	isProcessed = true;
+	return SQRESULT_NOTNULL;
+}
+
 static void CServerGameDLL__OnReceivedSayTextMessageHook(CServerGameDLL* self, unsigned int senderPlayerIndex, const char* text, int channelId) {
-	if (gSkipSayTextHook) {
-		gSkipSayTextHook = false;
+	// Set by ChatSendFromPlayer so it can bypass any checks
+	if (g_SkipSayTextHook) {
+		g_SkipSayTextHook = false;
 		CServerGameDLL__OnReceivedSayTextMessageUnhooked(self, senderPlayerIndex, text, channelId);
 		return;
 	}
 
-	// todo: impl other hook code here
+	void* sender = GetPlayerByIndex(senderPlayerIndex - 1);
 
-	CServerGameDLL__OnReceivedSayTextMessageUnhooked(self, senderPlayerIndex, text, channelId);
+	// check chat ratelimits
+	if (!g_ServerAuthenticationManager->CheckPlayerChatRatelimit(sender)) {
+		return;
+	}
+
+	bool shouldDoChathooks = strstr(GetCommandLineA(), "-enablechathooks");
+	if (shouldDoChathooks)
+	{
+
+		currentMessage = text;
+		currentPlayerId = senderPlayerIndex - 1; // Stupid fix cause of index offsets
+		currentChannelId = channelId;
+		shouldBlock = false;
+		isProcessed = false;
+
+		g_ServerSquirrelManager->ExecuteCode("CServerGameDLL_ProcessMessageStartThread()");
+		if (!shouldBlock && currentPlayerId + 1 == senderPlayerIndex) // stop player id spoofing from server
+		{
+			CServerGameDLL__OnReceivedSayTextMessage(self, currentPlayerId + 1, currentMessage.c_str(), currentChannelId);
+		}
+	}
+	else
+	{
+		CServerGameDLL__OnReceivedSayTextMessage(self, senderPlayerIndex, text, channelId);
+	}
 }
 
 ChatRichText::ChatRichText() {
@@ -107,7 +180,7 @@ void ChatRichText::SetFadeLengthSecs(float fadeLength) {
 void ChatSendFromPlayer(unsigned int playerId, const char* text, bool isteam) {
 	unsigned int playerIndex = playerId + 1;
 
-	gSkipSayTextHook = true;
+	g_SkipSayTextHook = true;
 	CServerGameDLL__OnReceivedSayTextMessage(gServer, playerIndex, text, isteam ? 1 : 0);
 }
 
@@ -224,6 +297,75 @@ void ChatWhisperRichText(unsigned int toPlayerId, const ChatRichText& text) {
 	CRecipientFilter__Destruct(&filter);
 }
 
+// void function NSChatSendFromPlayer( int playerIndex, string text, bool isteam )
+SQRESULT SQ_ChatSendFromPlayer(void* sqvm) {
+	int playerIndex = ServerSq_getinteger(sqvm, 1);
+	const char* text = ServerSq_getstring(sqvm, 2);
+	bool isteam = ServerSq_getbool(sqvm, 3);
+
+	ChatSendFromPlayer(playerIndex, text, isteam);
+
+	return SQRESULT_NULL;
+}
+
+// void function NSChatWhisperFromPlayer( int fromPlayerIndex, int toPlayerIndex, string text )
+SQRESULT SQ_ChatWhisperFromPlayer(void* sqvm) {
+	int fromPlayerIndex = ServerSq_getinteger(sqvm, 1);
+	int toPlayerIndex = ServerSq_getinteger(sqvm, 2);
+	const char* text = ServerSq_getstring(sqvm, 3);
+
+	ChatWhisperFromPlayer(fromPlayerIndex, toPlayerIndex, text);
+
+	return SQRESULT_NULL;
+}
+
+// void function NSChatBroadcastText( string text )
+SQRESULT SQ_ChatBroadcastText(void* sqvm) {
+	const char* text = ServerSq_getstring(sqvm, 1);
+
+	ChatBroadcastText(text);
+
+	return SQRESULT_NULL;
+}
+
+// void function NSChatBroadcastRainbowText( string text )
+SQRESULT SQ_ChatBroadcastRainbowText(void* sqvm) {
+	const char* text = ServerSq_getstring(sqvm, 1);
+
+	unsigned char colors[18];
+	colors[0] = 0xFF; colors[1] = 0x7F; colors[2] = 0x7F;
+	colors[3] = 0xFF; colors[4] = 0xFF; colors[5] = 0x7F;
+	colors[6] = 0x7F; colors[7] = 0xFF; colors[8] = 0x7F;
+	colors[9] = 0x7F; colors[10] = 0xFF; colors[11] = 0xFF;
+	colors[12] = 0x7F; colors[13] = 0x7F; colors[14] = 0xFF;
+	colors[15] = 0xFF; colors[16] = 0x7F; colors[17] = 0xFF;
+
+	ChatRichText richText;
+	size_t textLen = strlen(text);
+	for (size_t charIndex = 0; charIndex < textLen; charIndex++) {
+		size_t colorIndex = charIndex % 6;
+		richText.InsertColor(colors[colorIndex * 3], colors[colorIndex * 3 + 1], colors[colorIndex * 3 + 2], 255);
+
+		char mini[2];
+		mini[0] = text[charIndex];
+		mini[1] = 0;
+		richText.InsertText(mini);
+	}
+	ChatBroadcastRichText(richText);
+
+	return SQRESULT_NULL;
+}
+
+// void function NSChatWhisperText( int toPlayerIndex, string text )
+SQRESULT SQ_ChatWhisperText(void* sqvm) {
+	int toPlayerIndex = ServerSq_getinteger(sqvm, 1);
+	const char* text = ServerSq_getstring(sqvm, 2);
+
+	ChatWhisperText(toPlayerIndex, text);
+
+	return SQRESULT_NULL;
+}
+
 void InitialiseServerChatHooks_Engine(HMODULE baseAddress) {
 	gServer = (CServerGameDLL*)((char*)baseAddress + 0x13F0AA98);
 }
@@ -247,4 +389,19 @@ void InitialiseServerChatHooks_Server(HMODULE baseAddress) {
 	ENABLER_CREATEHOOK(
 		hook, (char*)baseAddress + 0x1595C0, &CServerGameDLL__OnReceivedSayTextMessageHook,
 		reinterpret_cast<LPVOID*>(&CServerGameDLL__OnReceivedSayTextMessage));
+
+	// Chat intercept functions
+	g_ServerSquirrelManager->AddFuncRegistration("void", "NSSetMessage", "string message, int playerId, int channelId, bool shouldBlock", "", setMessage);
+	g_ServerSquirrelManager->AddFuncRegistration("string", "NSChatGetCurrentMessage", "", "", getMessageServer);
+	g_ServerSquirrelManager->AddFuncRegistration("int", "NSChatGetCurrentPlayer", "", "", getPlayerServer);
+	g_ServerSquirrelManager->AddFuncRegistration("int", "NSChatGetCurrentChannel", "", "", getChannelServer);
+	g_ServerSquirrelManager->AddFuncRegistration("bool", "NSShouldProcessMessage", "", "", getShouldProcessMessage);
+	g_ServerSquirrelManager->AddFuncRegistration("void", "NSPushMessage", "string message, int playerId, int channelId, bool shouldBlock", "", pushMessage);
+
+	// Chat sending functions
+	g_ServerSquirrelManager->AddFuncRegistration("void", "NSChatSendFromPlayer", "int playerIndex, string text, bool isteam", "", SQ_ChatSendFromPlayer);
+	g_ServerSquirrelManager->AddFuncRegistration("void", "NSChatWhisperFromPlayer", "int fromPlayerIndex, int toPlayerIndex, string text", "", SQ_ChatWhisperFromPlayer);
+	g_ServerSquirrelManager->AddFuncRegistration("void", "NSChatBroadcastText", "string text", "", SQ_ChatBroadcastText);
+	g_ServerSquirrelManager->AddFuncRegistration("void", "NSChatBroadcastRainbowText", "string text", "", SQ_ChatBroadcastRainbowText);
+	g_ServerSquirrelManager->AddFuncRegistration("void", "NSChatWhisperText", "int toPlayerIndex, string text", "", SQ_ChatWhisperText);
 }
